feat(util): move classe util para pasta xml
